moving files out of docker_utils folders ahead of celery migration
