+++ backend/docker_utils/config.py
@@ -7,13 +7,13 @@ import os
 CONTAINERS = {
     "validator": {
         "port": "8001:8001",
-        "dockerfile": "docker_utils/Dockerfile",
+        "dockerfile": "backend/Dockerfile",
         "name": "validator",
         "service_file": "docker/services/validation_server.py",
     },
     "simulator": {
         "port": "8002:8002",
-        "dockerfile": "docker_utils/Dockerfile",
+        "dockerfile": "backend/Dockerfile",
         "name": "simulator",
         "service_file": "docker/services/simulation_server.py",
     },
