+++ config/deploy.yml
@@ -52,7 +52,7 @@ registry:
 
 builder:
   arch: amd64
-  dockerfile: backend/docker_utils/Dockerfile
+  dockerfile: backend/Dockerfile
   context: .
   args:
     RUNTIME_PACKAGES: postgresql18-client
